Fix buggy task when changing hostname
@ltalirz this is not catched by the molecule
test because this is run only for VMs and not on
docker. From the comments, apparently /etc/hosts cannot
be changed and this is the reason why this is skipped
(still, changing the hostname is something useful and some
things should be doable also inside docker).

This at least fixes the problem, but suggestions are
welcome on hot to test this with molecule