delete leftover from PBS scheduler
The PBS scheduler required the hostname to be associated with the
dynamic IP address in /etc/hosts.

Setting this via ansible is is problematic when the IP changes,
e.g. when creating new VMs from an existing image.
The slurm role does not require this line (and already has a built-in
mechanism for dealing with related issues).